lib/storage: move rawRowShards to raw_row.go
This way the related types are kept together: rawRow, rawRowsShard, and rawRowShards.

Next steps:
- Break the dependency on partition type by introducing rawRowsFlusher interface
- Write tests for rawRowsShard, and rawRowsShards and put them in raw_row_test.go
